# Conflicts:
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/config/SecurityConfig.java
#	swaggerDocumentation.yaml

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/interfaces/ImagesApi.java
@@ -63,7 +63,7 @@ public interface ImagesApi {
         produces = { "image/_*" }
     )
     default ResponseEntity<org.springframework.core.io.Resource> imagesIdGet(
-        @Parameter(name = "id", description = "ID of the image that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Integer id
+        @Parameter(name = "id", description = "ID of the image that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Long id
     ) {
         return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
